These are the initial files
1. Process Files: 
a. This file takes care of loading the  data files (S1.txt - S4.txt) found in the MQ2007 folder.
b. The data is loaded as a pandas data frame with feature numbers stripped out so we can use them as numeric values 
c. A small set of features are "enginered". Interaction of outlink and inlink is added by multiplying those columns 
d.  We add link information by simply concatenating the data from siteMap. (note dictionary is needed to translate keys correctly) 
-- getSiteFeats.py 
 a.  Adds the site feats as a pandas matrix 
--- featureNames.txt 
a. I named all the columns so they oculd be easily refered to